Refactor: QuestionInput 컴포넌트 통합
