--- src/features/profile/ui/inputs/index.ts
@@ -1,3 +1,0 @@
-export { TextInput } from "./TextInput";
-export { SelectInput } from "./SelectInput";
-export { NumberInput } from "./NumberInput";
